include the old mummy code (for a while anyway)
since there are backwards-incompatibilities with the new mummy code, go
ahead and include the old version as the module "oldmummy", so users can
"import oldmummy as mummy" and carry on.

the tests have also been modified to run the full test suite against
oldmummy, and new tests have been added to check that everything is
backwards-compatible that should be.